Fix mistake in test
Was using an old activator and was succeeding for the wrong reasons. There was some code in the activator which
registered an extension type called CANNOT_FIND_EXTENSION_TYPE, if the extension type could not be found. This code
was changed, but the old activator continued to be used. So, the test passed. But, with new activator code, it failed
to register the extension.