Fix inconsistent ordering when ordering by `lastTransitionTime`
The `buildStateTransition` table contains a `NULL` transition time for
`completed` state for jobs that are re-scheduled. This behavior by
itself is not a problem as such. However when sorting based on the
`lastTransitionTime`, `NULL` values need to be handled more gracefully
by adding a `NULLS FIRST` or `NULLS LAST` depending on the desired
ordering.

Fixes #7138