Remove duplicate x-axis labels.
The way the data was rendered into the graph was causing multiple labels
on the x-axis for stage reruns.

This commit also changes behavior of the library to get around
limitations of the library. We only show the "latest" passsing and
failing stage counters at each pipeline counter. Earlier, highcharts
would show every stage rerun on the graph (so you could have several
y-axis "dots" for each value on the x-axis (which shows the pipeline
counter)