Spark implementation for Status Reports
This uses a freemarker template for implementation rather than a JavaScript SPA. The Status Report pages come from the plugin, and include a page header, we can't use a JS SPA without changing the plugin view. The plugin view also relies on the css from the rails pages, which had to be included in the freemarker template.

At some point, we will have to change the plugin views to make it compatible with the new SPAs and remove the rails CSS.